refactor: creates ai/ directory
