+++ listeners/events/app_mentioned.py
@@ -2,7 +2,7 @@ from logging import Logger
 from slack_sdk import WebClient
 from slack_bolt import Say
 
-from ..llm_caller import call_llm
+from ai.llm_caller import call_llm
 from ..views.feedback_block import create_feedback_block
 
 """
